requirements.txt
fastapi
uvicorn[standard]
instagrapi
python-dotenv
selenium
webdriver-manager
beautifulsoup4
